small updates for better error handling :D

+++ app/dashboard/_actions/deleteBook.ts
@@ -4,6 +4,8 @@ import prisma from '@/app/lib/prisma';
 import { revalidatePath } from 'next/cache';
 
 export async function DeleteBook(id: string) {
+  if (!id) throw new Error('Book can not be identified');
+
   const deletedBook = await prisma.book.delete({
     where: {
       id,
@@ -13,6 +15,6 @@ export async function DeleteBook(id: string) {
   if (!deletedBook) {
     throw new Error('Book Not Found!');
   }
-  
+
   return deletedBook;
 }
